moved shutdown into main added tests and error reporter
