+++ main.go
@@ -4,11 +4,19 @@ import (
 	"os"
 	"strconv"
 
+	"fmt"
+	"github.com/ONSdigital/dp-dimension-extractor/api"
 	"github.com/ONSdigital/dp-dimension-extractor/config"
+	"github.com/ONSdigital/dp-dimension-extractor/event"
 	"github.com/ONSdigital/dp-dimension-extractor/service"
+	"github.com/ONSdigital/dp-reporter-client/reporter"
 	"github.com/ONSdigital/go-ns/kafka"
 	"github.com/ONSdigital/go-ns/log"
+	"github.com/ONSdigital/go-ns/rchttp"
 	"github.com/ONSdigital/go-ns/s3"
+	"golang.org/x/net/context"
+	"os/signal"
+	"syscall"
 )
 
 func main() {
@@ -44,18 +52,93 @@ func main() {
 		os.Exit(1)
 	}
 
-	svc := &service.Service{
+	dimensionExtractedErrProducer, err := kafka.NewProducer(cfg.Brokers, cfg.EventReporterTopic, int(envMax))
+	if err != nil {
+		log.Error(err, nil)
+		os.Exit(1)
+	}
+
+	signals := make(chan os.Signal, 1)
+	signal.Notify(signals, syscall.SIGINT, syscall.SIGTERM)
+
+	eventLoopDone := make(chan bool)
+	apiErrors := make(chan error, 1)
+
+	api.CreateDimensionExtractorAPI(cfg.DimensionExtractorURL, cfg.BindAddr, apiErrors)
+
+	service := &service.Service{
 		EnvMax:                envMax,
-		BindAddr:              cfg.BindAddr,
-		Consumer:              syncConsumerGroup,
 		DatasetAPIURL:         cfg.DatasetAPIURL,
 		DatasetAPIAuthToken:   cfg.DatasetAPIAuthToken,
 		DimensionExtractorURL: cfg.DimensionExtractorURL,
 		MaxRetries:            cfg.MaxRetries,
 		Producer:              dimensionExtractedProducer,
 		S3:                    s3,
-		Shutdown:              cfg.GracefulShutdownTimeout,
+		HTTPClient:            rchttp.DefaultClient,
 	}
 
-	svc.Start()
+	errorReporter, err := reporter.NewImportErrorReporter(dimensionExtractedErrProducer, log.Namespace)
+	if err != nil {
+		log.ErrorC("error while attempting to create error reporter client", err, nil)
+		os.Exit(1)
+	}
+
+	eventConsumer := event.Consumer{
+		KafkaConsumer: syncConsumerGroup,
+		EventService:  service,
+		ErrorReporter: errorReporter,
+	}
+
+	eventLoopContext, eventLoopCancel := context.WithCancel(context.Background())
+	eventConsumer.Start(eventLoopDone, eventLoopContext)
+
+	// block until a fatal error, signal or eventLoopDone - then proceed to shutdown
+	select {
+	case <-eventLoopDone:
+		log.Debug("quitting after done was closed", nil)
+	case signal := <-signals:
+		log.Debug("quitting after os signal received", log.Data{"signal": signal})
+	case consumerError := <-syncConsumerGroup.Errors():
+		log.Error(fmt.Errorf("aborting consumer"), log.Data{"message_received": consumerError})
+	case producerError := <-service.Producer.Errors():
+		log.Error(fmt.Errorf("aborting producer"), log.Data{"message_received": producerError})
+	case <-apiErrors:
+		log.Error(fmt.Errorf("server error forcing shutdown"), nil)
+	}
+
+	// give the app `Timeout` seconds to close gracefully before killing it.
+	ctx, cancel := context.WithTimeout(context.Background(), cfg.GracefulShutdownTimeout)
+
+	go func() {
+		log.Debug("stopping kafka consumer listener", nil)
+		syncConsumerGroup.StopListeningToConsumer(ctx)
+		log.Debug("stopped kafka consumer listener", nil)
+		eventLoopCancel()
+		<-eventLoopDone
+		log.Debug("closing http server", nil)
+		if err := api.Close(ctx); err != nil {
+			log.ErrorC("failed to gracefully close http server", err, nil)
+		} else {
+			log.Debug("gracefully closed http server", nil)
+		}
+		log.Debug("closing kafka producer", nil)
+		service.Producer.Close(ctx)
+		log.Debug("closed kafka producer", nil)
+		log.Debug("closing kafka consumer", nil)
+		syncConsumerGroup.Close(ctx)
+		log.Debug("closed kafka consumer", nil)
+
+		log.Info("done shutdown - cancelling timeout context", nil)
+
+		cancel() // stop timer
+	}()
+
+	// wait for timeout or success (via cancel)
+	<-ctx.Done()
+	if ctx.Err() == context.DeadlineExceeded {
+		log.Error(ctx.Err(), nil)
+	} else {
+		log.Info("done shutdown gracefully", log.Data{"context": ctx.Err()})
+	}
+	os.Exit(1)
 }
